fix(#18): QueryKey 재설정
