+++ src/components/auth/login/login-form.tsx
@@ -5,7 +5,7 @@ import AuthInputGroup from "@/components/auth/auth-input-group";
 import { Button } from "@/components/ui/button";
 import { FieldError } from "@/components/ui/field";
 import { Spinner } from "@/components/ui/spinner";
-import { LOGIN_PARAM, PROFILE_QUERY_KEY } from "@/constants/auth";
+import { LOGIN_PARAM, USER_QUERY_KEY } from "@/constants/auth";
 import { createClient } from "@/lib/supabase/client";
 import { cn } from "@/lib/utils";
 import { loginSchema } from "@/lib/zod/auth";
@@ -65,8 +65,9 @@ export default function LoginForm({ loading, onLoadingChange: setIsLoading }: Lo
     }
 
     setUser(authUser);
-    queryClient.invalidateQueries({ queryKey: PROFILE_QUERY_KEY });
+    await queryClient.invalidateQueries({ queryKey: USER_QUERY_KEY.all });
 
+    setIsLoading(null);
     router.push(`/${LOGIN_PARAM}`);
     router.refresh();
   };
