+++ src/components/setting/profile/profile-form.tsx
@@ -22,7 +22,7 @@ import { Spinner } from "@/components/ui/spinner";
 
 import { checkNicknameAction, updateProfileAction } from "@/actions/auth";
 import ProfileFormSkeleton from "@/components/setting/profile/profile-form-skeleton";
-import { PROFILE_QUERY_KEY } from "@/constants/auth";
+import { USER_QUERY_KEY } from "@/constants/auth";
 import { useUser } from "@/hooks/use-profile";
 import { cn } from "@/lib/utils";
 import { ProfileFormValues, profileSchema } from "@/lib/zod/auth";
@@ -148,11 +148,17 @@ export default function ProfileForm() {
       return;
     }
 
+    reset({
+      ...data,
+      photoUrl: result.photoUrl,
+    });
+
+    queryClient.invalidateQueries({ queryKey: USER_QUERY_KEY.db() });
+
     setIsSaving(false);
     setVerifiedNickname(data.nickname);
     setPendingFile(null);
 
-    queryClient.invalidateQueries({ queryKey: PROFILE_QUERY_KEY });
     toast.success("프로필 업데이트 완료");
   };
