+++ src/constants/footer.ts
@@ -1,0 +1,5 @@
+export const members = [
+  { name: "전지호", github: "wjswlgh96" },
+  { name: "이주영", github: "ele-003" },
+  { name: "안혜진", github: "Hyejinjin-An" },
+];
