+++ src/hooks/use-profile.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { PROFILE_QUERY_KEY } from "@/constants/auth";
+import { USER_QUERY_KEY } from "@/constants/auth";
 import { createClient } from "@/lib/supabase/client";
 import { useAuthStore } from "@/stores/auth";
 import { DBUser } from "@/types/user";
@@ -10,13 +10,12 @@ import { useQuery } from "@tanstack/react-query";
  * 현재 로그인된 유저의 public.user 프로필을 조회.
  * - AuthUser(Zustand)가 준비되면 자동 실행
  * - 5분간 캐싱 → 여러 컴포넌트에서 호출해도 네트워크 1회
- * - 프로필 업데이트 후에는 `queryClient.invalidateQueries({ queryKey: PROFILE_QUERY_KEY })`로 갱신
  */
 export function useUser() {
   const user = useAuthStore((s) => s.user);
 
   return useQuery<DBUser | null>({
-    queryKey: [...PROFILE_QUERY_KEY, user?.id],
+    queryKey: USER_QUERY_KEY.db(),
     queryFn: async () => {
       if (!user) return null;
       const supabase = createClient();
